Added more delete functions to all services

+++ Libraries/DigiBugzy.Services/Administration/Categories/CategoryCustomFieldService.cs
@@ -1,4 +1,6 @@
 ﻿
+using Microsoft.EntityFrameworkCore;
+
 namespace DigiBugzy.Services.Administration.Categories
 {
     public class CategoryCustomFieldService: BaseService, ICategoryCustomFieldService
@@ -23,11 +25,43 @@ namespace DigiBugzy.Services.Administration.Categories
         }
 
         /// <inheritdoc />
-        public void Delete(CategoryCustomField entity)
+        public void Delete(CategoryCustomField entity, bool hardDelete = true)
         {
-            dbContext.CategoryCustomFields.Remove(entity);
-            dbContext.SaveChanges();
+            if (hardDelete)
+            {
+                dbContext.CategoryCustomFields.Remove(entity);
+                dbContext.SaveChanges();
+            }
+            else
+            {
+                entity.IsDeleted = true;
+                entity.IsActive = false;
+                dbContext.CategoryCustomFields.Update(entity);
+                dbContext.SaveChanges();
+            }
         }
+
+        /// <inheritdoc />
+        public void Delete(List<CategoryCustomField> entities, bool hardDelete = true)
+        {
+            if (hardDelete)
+            {
+                dbContext.CategoryCustomFields.RemoveRange(entities);
+                dbContext.SaveChanges();
+            }
+            else
+            {
+                foreach (var entity in entities)
+                {
+                    entity.IsDeleted = true;
+                    entity.IsActive = false;
+                    dbContext.CategoryCustomFields.Update(entity);
+                    dbContext.SaveChanges();
+                }
+            }
+        }
+
+        
 
         #endregion
